Converted RNN to use tokenizer
